gateway/pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = ""
authors = ["Pumidol Leelerdsakulvong <pumidol.l@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-ninja = "^1.1.0"
whitenoise = "^6.6.0"
psycopg = {extras = ["binary"], version = "^3.1.14"}
pydantic = {extras = ["email"], version = "^2.5.2"}
django-cors-headers = "^4.3.1"
hypercorn = "^0.15.0"
rich = "^13.7.0"
celery = {extras = ["redis"], version = "^5.3.6"}
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
